Outdated portable installer x 1.
This commit is the first in a commit chain adding a Lutris installer for
so-called outdated portable variants of Aokana (i.e., distributions of
Aokana providing a portable unzipped post-installed `Aokana - Four
Rhythms Across the Blue/` directory containing the main `Aokana.exe`
Windows executable that is *not* the newest available official version).
Note this installer has yet to be rigorously tested.